Extract SSHKeyManager and ConfigManager from app.py, move source to src/
Split app.py (718 lines) into focused modules:
- src/ssh_keys.py: SSHKeyManager class for SSH key validation and storage
- src/config_manager.py: ConfigManager class for layered config form logic
- src/app.py: now ~300 lines of Flask routes and wiring

All .py files moved to src/ directory with pytest pythonpath configured.
