+++ src/app.py
@@ -1,0 +1,406 @@
+from flask import Flask, render_template, request, redirect, url_for, jsonify
+import os
+import subprocess
+
+from pydantic import ValidationError
+from config_schema import (
+    AdsbTruthConfig, Tar1090Config,
+    CaptureFormConfig, LocationFormConfig,
+)
+from config_manager import ConfigManager
+from device_state import DeviceState
+from form_utils import schema_to_form_fields
+from mender import MenderClient, get_latest_stable_from_github
+from ssh_keys import SSHKeyManager
+
+PROJECT_ROOT = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
+app = Flask(__name__, template_folder=os.path.join(PROJECT_ROOT, 'templates'))
+
+# Configurable paths - override via environment for local dev
+DATA_DIR = os.environ.get('DATA_DIR', '/data/retina-gui')
+USER_CONFIG_PATH = os.environ.get('USER_CONFIG_PATH', '/data/retina-node/config/user.yml')
+MERGED_CONFIG_PATH = os.environ.get('MERGED_CONFIG_PATH', '/data/retina-node/config/config.yml')
+RETINA_NODE_PATH = os.environ.get('RETINA_NODE_PATH', '/data/mender-app/retina-node/manifests')
+NODE_ID_FILE = os.environ.get('NODE_ID_FILE', '/data/mender/node_id')
+
+# SSH key manager
+ssh_keys = SSHKeyManager(os.path.join(DATA_DIR, "authorized_keys"))
+
+# Configuration manager
+config_mgr = ConfigManager(
+    user_config_path=USER_CONFIG_PATH,
+    merged_config_path=MERGED_CONFIG_PATH,
+    retina_node_path=RETINA_NODE_PATH,
+)
+
+# Mender client for device-initiated OTA
+mender = MenderClient(
+    server_url=os.environ.get('MENDER_SERVER_URL', 'https://hosted.mender.io'),
+    release_name=os.environ.get('MENDER_RELEASE_NAME', 'retina-node'),
+    device_type=os.environ.get('MENDER_DEVICE_TYPE', 'pi5-v3-arm64'),
+)
+
+# Device state machine — manages install locks, cloud services, and update detection
+MENDER_SERVICES = ["mender-authd", "mender-updated", "mender-connect"]
+
+device_state = DeviceState(
+    data_dir=DATA_DIR,
+    mender_services=MENDER_SERVICES,
+    mender_conf_path="/data/mender/mender.conf",
+    mender_conf_backup_dir="/data/mender-cloud-disabled",
+    mender_conf_backup_path="/data/mender-cloud-disabled/mender.conf",
+)
+
+# Enforce cloud services preference on app startup
+device_state.apply_startup_preferences()
+
+
+def get_node_id():
+    """Get node_id from Mender device identity file.
+
+    Reads from /data/mender/node_id which is generated by the
+    mender-device-identity script on first boot.
+
+    Returns node_id string or 'Unknown' if not available.
+    """
+    try:
+        with open(NODE_ID_FILE, 'r') as f:
+            node_id = f.read().strip()
+            if node_id:
+                return node_id
+    except FileNotFoundError:
+        app.logger.debug(f"Node ID file not found: {NODE_ID_FILE}")
+    except Exception as e:
+        app.logger.warning(f"Could not read node_id from {NODE_ID_FILE}: {e}")
+    return 'Unknown'
+
+
+@app.route("/")
+def index():
+    """Home page with node ID, services, and SSH keys."""
+    keys = ssh_keys.get_keys()
+
+    # Get node_id from Mender device identity file
+    node_id = get_node_id()
+
+    # Get software versions from Mender
+    owl_os_version, retina_node_version = mender.get_versions()
+
+    return render_template("index.html",
+                           ssh_keys=keys,
+                           node_id=node_id,
+                           owl_os_version=owl_os_version,
+                           retina_node_version=retina_node_version)
+
+
+@app.route("/eula")
+def eula():
+    """Display EULA page."""
+    return render_template("eula.html")
+
+
+@app.route("/config")
+def config_page():
+    """Configuration page with all settings.
+
+    Reads from merged config.yml to display actual running values.
+    """
+    config = config_mgr.load_merged_config()
+    retina_installed = config_mgr.is_retina_node_installed()
+
+    # Generate form fields from Pydantic schemas
+    capture_flat = ConfigManager.flatten_capture_for_form(config.get('capture', {}))
+    capture_fields = schema_to_form_fields(CaptureFormConfig, capture_flat)
+
+    location_flat = ConfigManager.flatten_location_for_form(config.get('location', {}))
+    location_fields = schema_to_form_fields(LocationFormConfig, location_flat)
+
+    truth_adsb_values = (config.get('truth', {}) or {}).get('adsb', {}) or {}
+    truth_fields = schema_to_form_fields(AdsbTruthConfig, truth_adsb_values)
+
+    # tar1090 needs special handling for adsb_source split
+    tar1090_values = ConfigManager.parse_tar1090_adsb_source(config)
+    tar1090_fields = schema_to_form_fields(Tar1090Config, tar1090_values)
+
+    return render_template("config.html",
+                           retina_installed=retina_installed,
+                           capture_fields=capture_fields,
+                           location_fields=location_fields,
+                           truth_fields=truth_fields,
+                           tar1090_fields=tar1090_fields)
+
+
+@app.route("/ssh-keys", methods=["POST"])
+def add_key():
+    key = request.form.get("ssh_key", "").strip()
+    if key and SSHKeyManager.is_valid_ssh_key(key):
+        ssh_keys.add_key(key)
+        return redirect(url_for("index"))
+    else:
+        keys = ssh_keys.get_keys()
+        return render_template("index.html", ssh_keys=keys,
+                               error="Invalid SSH key format")
+
+
+@app.route("/ssh-keys/delete", methods=["POST"])
+def delete_key():
+    key = request.form.get("ssh_key", "")
+    if key:
+        ssh_keys.remove_key(key)
+    return redirect(url_for("index"))
+
+
+@app.route("/config/save", methods=["POST"])
+def save_config():
+    """Save config form data to user.yml.
+
+    Only writes values that differ from the merged config.
+    This keeps user.yml clean with just the user's explicit overrides.
+    """
+    # Parse flat form data into section dicts
+    capture_flat, location_flat, truth_data, tar1090_data = ConfigManager.parse_flat_form_data(request.form.to_dict())
+
+    # Collect all validation errors
+    all_errors = {}
+
+    # Validate capture (using flat schema)
+    if capture_flat:
+        try:
+            CaptureFormConfig(**capture_flat)
+        except ValidationError as e:
+            all_errors.update(ConfigManager.format_validation_errors(e, 'capture'))
+
+    # Validate location (using flat schema)
+    if location_flat:
+        try:
+            LocationFormConfig(**location_flat)
+        except ValidationError as e:
+            all_errors.update(ConfigManager.format_validation_errors(e, 'location'))
+
+    # Validate truth (still nested)
+    if truth_data:
+        try:
+            AdsbTruthConfig(**truth_data)
+        except ValidationError as e:
+            all_errors.update(ConfigManager.format_validation_errors(e, 'truth'))
+
+    # Validate tar1090
+    if tar1090_data:
+        try:
+            Tar1090Config(**tar1090_data)
+        except ValidationError as e:
+            all_errors.update(ConfigManager.format_validation_errors(e, 'tar1090'))
+
+    # If validation errors, re-render form with flat data
+    if all_errors:
+        return render_template("config.html",
+                               retina_installed=config_mgr.is_retina_node_installed(),
+                               capture_fields=schema_to_form_fields(CaptureFormConfig, capture_flat),
+                               location_fields=schema_to_form_fields(LocationFormConfig, location_flat),
+                               truth_fields=schema_to_form_fields(AdsbTruthConfig, truth_data),
+                               tar1090_fields=schema_to_form_fields(Tar1090Config, tar1090_data),
+                               config_errors=all_errors)
+
+    # Convert flat form data to nested YAML structure
+    capture_nested = ConfigManager.unflatten_capture_from_form(capture_flat)
+    location_nested = ConfigManager.unflatten_location_from_form(location_flat)
+
+    # Join tar1090 adsb_source fields
+    tar1090_nested = {}
+    if tar1090_data:
+        host = tar1090_data.pop('adsb_source_host', '')
+        port = tar1090_data.pop('adsb_source_port', '')
+        protocol = tar1090_data.pop('adsb_source_protocol', '')
+        if host or port or protocol:
+            tar1090_nested['adsb_source'] = f"{host},{port},{protocol}"
+        tar1090_nested.update(tar1090_data)
+
+    # Load merged config (for comparison) and existing user config
+    merged_config = config_mgr.load_merged_config()
+    existing_user = config_mgr.load_user_config()
+
+    # Compute overrides - only save values that differ from merged config
+    new_user_config = {}
+
+    # Preserve fields not managed by this form (e.g., network.node_id, process, save)
+    for key in existing_user:
+        if key not in ('capture', 'location', 'truth', 'tar1090'):
+            new_user_config[key] = existing_user[key]
+
+    # Compute and add overrides for each section
+    if capture_flat:
+        capture_overrides = config_mgr.compute_user_overrides(capture_nested, merged_config, existing_user, 'capture')
+        if capture_overrides:
+            new_user_config['capture'] = capture_overrides
+
+    if location_flat:
+        location_overrides = config_mgr.compute_user_overrides(location_nested, merged_config, existing_user, 'location')
+        if location_overrides:
+            new_user_config['location'] = location_overrides
+
+    if truth_data:
+        truth_nested = {'adsb': truth_data}
+        truth_overrides = config_mgr.compute_user_overrides(truth_nested, merged_config, existing_user, 'truth')
+        if truth_overrides:
+            new_user_config['truth'] = truth_overrides
+
+    if tar1090_nested:
+        tar1090_overrides = config_mgr.compute_user_overrides(tar1090_nested, merged_config, existing_user, 'tar1090')
+        if tar1090_overrides:
+            new_user_config['tar1090'] = tar1090_overrides
+
+    config_mgr.save_user_config(new_user_config)
+    return redirect(url_for("config_page") + "?saved=1")
+
+
+@app.route("/config/apply", methods=["POST"])
+def apply_config():
+    """Run config-merger and restart services."""
+    if not config_mgr.is_retina_node_installed():
+        return jsonify({"success": False, "error": "retina-node not installed"}), 400
+
+    try:
+        # Run config-merger to merge user.yml with defaults
+        result = subprocess.run(
+            ["docker", "compose", "-p", "retina-node", "run", "--rm", "config-merger"],
+            cwd=RETINA_NODE_PATH,
+            capture_output=True,
+            text=True,
+            timeout=60
+        )
+        if result.returncode != 0:
+            return jsonify({
+                "success": False,
+                "error": f"config-merger failed: {result.stderr or result.stdout}"
+            }), 500
+
+        # Restart services with new config
+        result = subprocess.run(
+            ["docker", "compose", "-p", "retina-node", "up", "-d", "--force-recreate"],
+            cwd=RETINA_NODE_PATH,
+            capture_output=True,
+            text=True,
+            timeout=120
+        )
+        if result.returncode != 0:
+            return jsonify({
+                "success": False,
+                "error": f"restart failed: {result.stderr or result.stdout}"
+            }), 500
+
+        return jsonify({"success": True})
+
+    except subprocess.TimeoutExpired:
+        return jsonify({"success": False, "error": "Command timed out"}), 500
+    except Exception as e:
+        return jsonify({"success": False, "error": str(e)}), 500
+
+
+@app.route("/mender/check")
+def mender_check():
+    """Check for available updates and install status."""
+    _, current = mender.get_versions()
+
+    # Check if any update in progress (GUI install or server-pushed)
+    in_progress, reason = device_state.is_any_update_in_progress()
+    if in_progress:
+        locked, lock_info = device_state.is_install_locked()
+        return jsonify({
+            "installing": True,
+            "version": lock_info["version"] if lock_info else "system update",
+            "started_at": lock_info["started_at"] if lock_info else None,
+            "reason": reason
+        })
+
+    latest, error = get_latest_stable_from_github()
+    if error:
+        return jsonify({"error": error})
+
+    return jsonify({
+        "installing": False,
+        "latest_version": latest,
+        "current_version": current,
+        "update_available": current is None or latest != current
+    })
+
+
+@app.route("/mender/install", methods=["POST"])
+def mender_install():
+    """Install latest stable retina-node artifact from Mender."""
+    # Enable cloud services and wait for Mender auth
+    success, error = device_state.ensure_cloud_services_enabled(mender.get_jwt)
+    if not success:
+        return jsonify({"success": False, "error": error})
+
+    # Check if already installed
+    _, retina_node_version = mender.get_versions()
+    if retina_node_version:
+        return jsonify({"success": False, "error": "Already installed"})
+
+    # Guard: block if any update is already in progress
+    can_install, reason = device_state.can_start_install()
+    if not can_install:
+        return jsonify({"success": False, "error": reason}), 409
+
+    # Get latest stable version from GitHub releases
+    version_tag, error = get_latest_stable_from_github()
+    if error:
+        return jsonify({"success": False, "error": f"Failed to get version: {error}"})
+
+    # Acquire lock
+    release_name = f"retina-node-{version_tag}"
+    if not device_state.acquire_install_lock(release_name):
+        return jsonify({"success": False, "error": "Install already in progress"}), 409
+
+    try:
+        # Query Mender for that specific release
+        artifacts, error = mender.list_artifacts(release_name=release_name)
+        if error:
+            return jsonify({"success": False, "error": error})
+
+        if not artifacts:
+            return jsonify({"success": False, "error": f"No artifact found for {release_name}"})
+
+        artifact = artifacts[0]  # Should be exactly one for this release/device
+
+        # Get download URL
+        url, error = mender.get_download_url(artifact["id"])
+        if error:
+            return jsonify({"success": False, "error": error})
+
+        # Install
+        success, error = mender.install_from_url(url)
+        if not success:
+            return jsonify({"success": False, "error": error})
+
+        return jsonify({"success": True, "installed_version": release_name})
+    finally:
+        device_state.release_install_lock()
+
+
+@app.route("/mender/cloud-services", methods=["GET"])
+def cloud_services_status():
+    """Check if Mender cloud services are enabled."""
+    return jsonify(device_state.get_cloud_services_status())
+
+
+@app.route("/mender/cloud-services", methods=["POST"])
+def cloud_services_toggle():
+    """Enable or disable Mender cloud services."""
+    data = request.get_json()
+    if not data or "enabled" not in data:
+        return jsonify({"success": False, "error": "Missing 'enabled' field"}), 400
+
+    enabled = data["enabled"]
+
+    success, error = device_state.set_cloud_services(enabled)
+    if not success:
+        return jsonify({"success": False, "error": error}), 409
+
+    return jsonify({"success": True})
+
+
+if __name__ == "__main__":
+    port = int(os.environ.get('PORT', 80))
+    debug = os.environ.get('FLASK_DEBUG', 'false').lower() == 'true'
+    app.run(host="::", port=port, debug=debug)
